requirements.txt
# Core runtime
numpy>=1.26
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66

# Deep learning
torch==2.1.2
torchvision==0.16.2

# Metrics / vision utilities
lpips>=0.1.4
opencv-python-headless>=4.9
scikit-image>=0.23

# Offline caption generation
transformers==4.37.2
accelerate>=0.33
